+++ src/main/java/Observable.java
@@ -3,5 +3,9 @@
  * Only responsibility is to tell who will notify about state changes.
  */
 public interface Observable {
-    Notifier getNotifier();
+    /**
+     * Gives you the {@link Notifier} for the {@link Observable}. It returns {@link NotifiesObservers} to
+     * only expose that part of the interface that the {@link Observer} needs.
+     */
+    NotifiesObservers getNotifier();
 }