+++ src/main/java/WatchesObservable.java
@@ -1,0 +1,6 @@
+/**
+ * Part of the {@link Notifier} interface that {@link Observable} should use
+ */
+public interface WatchesObservable {
+    void notifyObservers();
+}